refactor: extract pure helpers from god hooks (chats fuzzySearch, IE display/url)
Behavior-neutral moves of pure, dependency-free helpers out of two large hooks,
plus de-duplication of real copy-paste.

- src/apps/chats/utils/fuzzySearch.ts: stripDiacritics/normalizeSearchText/
  isLooseSubsequence/levenshtein/bestSubstring/computeMatchScore/deriveScoreThreshold.
  useAiChat.ts and karaokeHandler.ts now import these instead of each defining
  their own identical copies (karaokeHandler had a ~120-line duplicate with a TODO).
- src/apps/internet-explorer/utils/displayNames.ts: getLanguageDisplayName/
  getLocationDisplayName.
- src/apps/internet-explorer/utils/urlHelpers.ts: MAX_TITLE_LENGTH/getHostnameFromUrl/
  formatTitle/decodeData/normalizeUrlForHistory.

No public hook return types change; only implementations moved + re-imported.
Verified: build, new fuzzySearch unit test, and IE renders correctly (window title
'Apple' is produced by the extracted formatTitle/getHostnameFromUrl).
